Do calculation with 64 bit precision to avoid overflow

+++ whisper.cpp
@@ -3387,7 +3387,7 @@ static int timestamp_to_sample(int64_t t, int n_samples) {
     return std::max(0, std::min((int) n_samples - 1, (int) ((t*WHISPER_SAMPLE_RATE)/100)));
 }
 
-static int64_t sample_to_timestamp(int i_sample) {
+static int64_t sample_to_timestamp(int64_t i_sample) {
     return (100*i_sample)/WHISPER_SAMPLE_RATE;
 }
